requirements.txt
celery==3.1.16
django-celery==3.1.16
django-simple-history==1.6.0
djangorestframework==3.1.1
Django==1.8
Markdown==2.6.1
six==1.9.0
